apps/Test_IA_CANDIDATURE/requirements.txt
langchain>=0.3.0
langchain-anthropic>=0.3.0
langgraph>=0.2.0
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.14.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
openai>=1.0.0
